refactor: convert dll-injector to standalone CLI binary
