+++ native/dll-injector/Cargo.toml
@@ -1,19 +1,13 @@
-
 [package]
 edition = "2021"
-name = "dll_injector"
+name = "dll-injector"
 version = "0.1.0"
 
-[lib]
-crate-type = ["cdylib"]
-
 [dependencies]
-napi = { version = "3.0.0", features = ["napi4"] }
-napi-derive = "3.0.0"
 libloading = "0.8"
 sysinfo = "0.30"
 anyhow = "1.0"
-
+clap = { version = "4.4.18", features = ["derive"] }
 
 [dependencies.windows]
 version = "0.52"
@@ -28,9 +22,10 @@ features = [
     "Win32_System_Diagnostics_ToolHelp",
 ]
 
-[build-dependencies]
-napi-build = "2"
-
 [profile.release]
 lto = true
 strip = "symbols"
+
+[[bin]]
+name = "dll-injector"
+path = "src/main.rs"
