+++ native/dll-injector/package.json
@@ -3,16 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
-  "napi": {
-    "binaryName": "dll-injector",
-    "targets": [
-      "x86_64-pc-windows-msvc"
-    ]
-  },
   "scripts": {
-    "build": "napi build --platform --release"
-  },
-  "devDependencies": {
-    "@napi-rs/cli": "^3.2.0"
+    "build": "cargo build --release"
   }
 }
